requirements.txt
absl-py==0.11.0
alembic==1.6.5
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.2.1
attrs==18.2.0
audioread==2.1.9
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.1
certifi==2018.11.29
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.4
entrypoints==0.3
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
fsspec==2021.7.0
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.26.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
gspread==3.7.0
gym==0.18.0
h5py==2.10.0
httplib2==0.19.1
huggingface-hub==0.0.14
idna==2.8
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.1
line-bot-sdk==1.19.0
llvmlite==0.37.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==4.3.0
multiprocess==0.70.12.2
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.2
newspaper3k==0.2.8
nltk==3.5
notebook==6.2.0
numba==0.54.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
OpenCC==1.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.8.0
pooch==1.5.2
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
py==1.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.7.4
pyimgur==0.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==4.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytrends==4.7.3
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.21.0
requests-file==1.5.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
scikit-learn==0.24.2
scipy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.12.0
smart-open==4.2.0
SoundFile==0.10.3.post1
soupsieve==2.2
SQLAlchemy==1.4.22
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.2.0
times==0.7
tinysegmenter==0.3
tldextract==3.1.0
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.24.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xxhash==2.0.2
